Added size limit and support for manual accept.
The SettingsPane has the Settings for the automaticDownloadLimit and the
MulticastSharer won't start a download when the filesize is bigger than
this settings. The event onAvailable is fired and the method accept()
exists to accept an file that is available.

Laks correct implementation, test now is accepting with a delay of 5
seconds.